feat(user):同一个账号仅允许在一台终端设备登录

+++ apps/authentication/signals_handlers.py
@@ -1,6 +1,7 @@
 from importlib import import_module
 
 from django.conf import settings
+from django.contrib.auth import user_logged_in
 from django.core.cache import cache
 from django.dispatch import receiver
 
@@ -9,14 +10,15 @@ from jms_oidc_rp.signals import openid_user_login_failed, openid_user_login_succ
 from .signals import post_auth_success, post_auth_failed
 
 
+@receiver(user_logged_in)
 def on_user_auth_login_success(sender, user, request, **kwargs):
-    # 仅允许单机登录
-    user_id = 'single_machine_login_' + str(user.id)
-    session_key = cache.get(user_id)
-    if session_key and session_key != request.session.session_key:
-        session = import_module(settings.SESSION_ENGINE).SessionStore(session_key)
-        session.delete()
-    cache.set(user_id, request.session.session_key)
+    if settings.USER_LOGIN_SINGLE_MACHINE_ENABLED:
+        user_id = 'single_machine_login_' + str(user.id)
+        session_key = cache.get(user_id)
+        if session_key and session_key != request.session.session_key:
+            session = import_module(settings.SESSION_ENGINE).SessionStore(session_key)
+            session.delete()
+        cache.set(user_id, request.session.session_key)
 
 
 @receiver(openid_user_login_success)
